requirements.txt
requests~=2.24.0
numpy~=1.18.5
vpython~=7.6.1
Pillow~=7.2.0
matplotlib~=3.2.2
scikit-image~=0.17.2
scipy~=1.4.1
tensorflow~=2.3.0
pandas~=1.1.1
pycocotools~=2.0
opencv-python~=4.3.0.36
trimesh~=3.8.6
neuralgym~=0.0.1